for box revenge precisely

+++ src/CDC/Collision.java
@@ -62,8 +62,8 @@ public class Collision {
     }
 
     // for box revenge
-    public static boolean isCollison(int p1LocX, int p1LocY,
-            int p1Size, ClientPlayerFeature player2) {
+    public static boolean isCollison(int p1LocX, int p1LocY, double p1Size,
+            ClientPlayerFeature player2) {
         double size = ((PlayerSize + p1Size) * 1.2);
         System.out.println(size + " " + (PlayerSize + p1Size) * 1.2);
         // this may change to not either 內切圓 外切圓 should be 混和圓
